app/requirements.txt
dash
dash-bootstrap-components
pandas
joblib
mysql-connector-python
flask
flask-login
passlib
fastapi
pydantic
uvicorn[standard]
pymysql
jwt
pylance
